Branch id (#60)
* update version

* update README

* update example README

* comment few tests for the moment

* comment few tests for the moment

* fix small bugs

* small changes: id hash added, some getters

* replace vector with map in Configuration

* branch id from std::hash

* replace uint

* replace int with size_t for id everywhere

* replace int with size_t for id everywhere

+++ infra/Cuts.hpp
@@ -58,7 +58,7 @@ class Cuts {
   }
 
   template<class A, class B>
-  bool Apply(const A& a, int a_id, const B& b, int b_id) const {
+  bool Apply(const A& a, size_t a_id, const B& b, size_t b_id) const {
     if (!is_init_) {
       throw std::runtime_error("Cuts::Apply - cut is not initialized!!");
     }
